pyproject.toml
[project]
name = "ml-pipeline-deployment"
version = "0.1.0"
description = "End-to-end ML pipeline deployment system for homelab infrastructure"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "boto3>=1.28.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "prometheus-client>=0.17.0",
    "mlflow>=2.8.0",
    "xgboost>=2.0.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.82.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
sagemaker = [
    "sagemaker>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["services*", "shared*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
